+++ flask_dance/consumer/oauth2.py
@@ -1,4 +1,3 @@
-from __future__ import unicode_literals, print_function
 import json
 
 import logging
@@ -55,7 +54,7 @@ class OAuth2ConsumerBlueprint(BaseOAuthConsumerBlueprint):
         session_class=None,
         storage=None,
         rule_kwargs=None,
-        **kwargs
+        **kwargs,
     ):
         """
         Most of the constructor arguments are forwarded either to the
@@ -183,7 +182,7 @@ class OAuth2ConsumerBlueprint(BaseOAuthConsumerBlueprint):
             state=self.state,
             blueprint=self,
             base_url=self.base_url,
-            **self.kwargs
+            **self.kwargs,
         )
 
         def token_updater(token):
@@ -204,7 +203,7 @@ class OAuth2ConsumerBlueprint(BaseOAuthConsumerBlueprint):
         url, state = self.session.authorization_url(
             self.authorization_url, state=self.state, **self.authorization_url_params
         )
-        state_key = "{bp.name}_oauth_state".format(bp=self)
+        state_key = f"{self.name}_oauth_state"
         flask.session[state_key] = state
         log.debug("state = %s", state)
         log.debug("redirect URL = %s", url)
@@ -241,7 +240,7 @@ class OAuth2ConsumerBlueprint(BaseOAuthConsumerBlueprint):
             )
             return redirect(next_url)
 
-        state_key = "{bp.name}_oauth_state".format(bp=self)
+        state_key = f"{self.name}_oauth_state"
         if state_key not in flask.session:
             # can't validate state, so redirect back to login view
             log.info("state not found, redirecting user to login")
@@ -261,7 +260,7 @@ class OAuth2ConsumerBlueprint(BaseOAuthConsumerBlueprint):
                 self.token_url,
                 authorization_response=request.url,
                 client_secret=self.client_secret,
-                **self.token_url_params
+                **self.token_url_params,
             )
         except MissingCodeError as e:
             e.args = (