+++ flask_dance/contrib/atlassian.py
@@ -5,82 +5,84 @@ from flask.globals import LocalProxy, _lookup_app_object
 from flask import _app_ctx_stack as stack
 
 
-__maintainer__ = "Pieter Ennes <support@authentiq.com>"
+__maintainer__ = "Przemyslaw Kanach <kanach16@gmail.com>"
 
 
-def make_authentiq_blueprint(
+def make_atlassian_blueprint(
     client_id=None,
     client_secret=None,
-    scope="openid profile",
+    scope=None,
+    reprompt_consent=False,
     redirect_url=None,
     redirect_to=None,
     login_url=None,
     authorized_url=None,
     session_class=None,
     storage=None,
-    hostname="connect.authentiq.io",
-    rule_kwargs=None,
 ):
     """
-    Make a blueprint for authenticating with authentiq using OAuth 2. This requires
-    a client ID and client secret from authentiq. You should either pass them to
+    Make a blueprint for authenticating with Atlassian using OAuth 2. This requires
+    a client ID and client secret from Atlassian. You should either pass them to
     this constructor, or make sure that your Flask application config defines
-    them, using the variables :envvar:`AUTHENTIQ_OAUTH_CLIENT_ID` and
-    :envvar:`AUTHENTIQ_OAUTH_CLIENT_SECRET`.
+    them, using the variables :envvar:`ATLASSIAN_OAUTH_CLIENT_ID` and
+    :envvar:`ATLASSIAN_OAUTH_CLIENT_SECRET`.
 
     Args:
-        client_id (str): The client ID for your application on Authentiq.
-        client_secret (str): The client secret for your application on Authentiq.
+        client_id (str): The client ID for your application on Atlassian.
+        client_secret (str): The client secret for your application on Atlassian.
         scope (str, optional): comma-separated list of scopes for the OAuth token.
+        reprompt_consent (bool): If True, force Atlassian to re-prompt the user
+            for their consent, even if the user has already given their
+            consent. Defaults to False.
         redirect_url (str): the URL to redirect to after the authentication
             dance is complete.
         redirect_to (str): if ``redirect_url`` is not defined, the name of the
             view to redirect to after the authentication dance is complete.
             The actual URL will be determined by :func:`flask.url_for`.
         login_url (str, optional): the URL path for the ``login`` view.
-            Defaults to ``/authentiq``.
+            Defaults to ``/atlassian``.
         authorized_url (str, optional): the URL path for the ``authorized`` view.
-            Defaults to ``/authentiq/authorized``.
+            Defaults to ``/atlassian/authorized``.
         session_class (class, optional): The class to use for creating a
             Requests session. Defaults to
             :class:`~flask_dance.consumer.requests.OAuth2Session`.
         storage: A token storage class, or an instance of a token storage
                 class, to use for this blueprint. Defaults to
                 :class:`~flask_dance.consumer.storage.session.SessionStorage`.
-        hostname (str, optional): If using a private instance of authentiq CE/EE,
-            specify the hostname, default is ``connect.authentiq.io``.
-        rule_kwargs (dict, optional): Additional arguments that should be passed when adding
-            the login and authorized routes. Defaults to ``None``.
 
     :rtype: :class:`~flask_dance.consumer.OAuth2ConsumerBlueprint`
     :returns: A :ref:`blueprint <flask:blueprints>` to attach to your Flask app.
     """
-    authentiq_bp = OAuth2ConsumerBlueprint(
-        "authentiq",
+    authorization_url_params = {"audience": "api.atlassian.com"}
+    if reprompt_consent:
+        authorization_url_params["prompt"] = "consent"
+
+    atlassian_bp = OAuth2ConsumerBlueprint(
+        "atlassian",
         __name__,
         client_id=client_id,
         client_secret=client_secret,
         scope=scope,
-        base_url=f"https://{hostname}/",
-        authorization_url=f"https://{hostname}/authorize",
-        token_url=f"https://{hostname}/token",
+        base_url="https://api.atlassian.com/",
+        authorization_url="https://auth.atlassian.com/authorize",
+        token_url="https://auth.atlassian.com/oauth/token",
         redirect_url=redirect_url,
         redirect_to=redirect_to,
         login_url=login_url,
         authorized_url=authorized_url,
+        authorization_url_params=authorization_url_params,
         session_class=session_class,
         storage=storage,
-        rule_kwargs=rule_kwargs,
     )
-    authentiq_bp.from_config["client_id"] = "AUTHENTIQ_OAUTH_CLIENT_ID"
-    authentiq_bp.from_config["client_secret"] = "AUTHENTIQ_OAUTH_CLIENT_SECRET"
+    atlassian_bp.from_config["client_id"] = "ATLASSIAN_OAUTH_CLIENT_ID"
+    atlassian_bp.from_config["client_secret"] = "ATLASSIAN_OAUTH_CLIENT_SECRET"
 
-    @authentiq_bp.before_app_request
+    @atlassian_bp.before_app_request
     def set_applocal_session():
         ctx = stack.top
-        ctx.authentiq_oauth = authentiq_bp.session
+        ctx.atlassian_oauth = atlassian_bp.session
 
-    return authentiq_bp
+    return atlassian_bp
 
 
-authentiq = LocalProxy(partial(_lookup_app_object, "authentiq_oauth"))
+atlassian = LocalProxy(partial(_lookup_app_object, "atlassian_oauth"))