+++ flask_dance/contrib/azure.py
@@ -1,13 +1,8 @@
-from __future__ import unicode_literals
-
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 from functools import partial
 from flask.globals import LocalProxy, _lookup_app_object
 
-try:
-    from flask import _app_ctx_stack as stack
-except ImportError:
-    from flask import _request_ctx_stack as stack
+from flask import _app_ctx_stack as stack
 
 
 __maintainer__ = "Steven MARTINS <steven.martins.fr@gmail.com>"
@@ -96,12 +91,8 @@ def make_azure_blueprint(
         client_secret=client_secret,
         scope=scope,
         base_url="https://graph.microsoft.com",
-        authorization_url="https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize".format(
-            tenant=tenant
-        ),
-        token_url="https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token".format(
-            tenant=tenant
-        ),
+        authorization_url=f"https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize",
+        token_url=f"https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token",
         redirect_url=redirect_url,
         redirect_to=redirect_to,
         login_url=login_url,