+++ flask_dance/contrib/discord.py
@@ -1,13 +1,8 @@
-from __future__ import unicode_literals
-
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 from functools import partial
 from flask.globals import LocalProxy, _lookup_app_object
 
-try:
-    from flask import _app_ctx_stack as stack
-except ImportError:
-    from flask import _request_ctx_stack as stack
+from flask import _app_ctx_stack as stack
 
 
 __maintainer__ = "Michael Delpech <michaeldel@protonmail.com>"
@@ -23,6 +18,7 @@ def make_discord_blueprint(
     authorized_url=None,
     session_class=None,
     storage=None,
+    prompt="consent",
     rule_kwargs=None,
 ):
     """
@@ -51,6 +47,9 @@ def make_discord_blueprint(
         storage: A token storage class, or an instance of a token storage
                 class, to use for this blueprint. Defaults to
                 :class:`~flask_dance.consumer.storage.session.SessionStorage`.
+        prompt (str, optional): Define authorization flow.
+            Defaults to ``consent``, setting it to ``None`` will skip user
+            interaction if the application was previously approved.
         rule_kwargs (dict, optional): Additional arguments that should be passed when adding
             the login and authorized routes. Defaults to `None.
 
@@ -58,6 +57,9 @@ def make_discord_blueprint(
     :returns: A :ref:`blueprint <flask:blueprints>` to attach to your Flask app.
     """
     scope = scope or ["identify"]
+    authorization_url_params = {"prompt": "consent"}
+    if prompt is None:
+        authorization_url_params["prompt"] = None
     discord_bp = OAuth2ConsumerBlueprint(
         "discord",
         __name__,
@@ -73,6 +75,7 @@ def make_discord_blueprint(
         authorized_url=authorized_url,
         session_class=session_class,
         storage=storage,
+        authorization_url_params=authorization_url_params,
         rule_kwargs=rule_kwargs,
     )
     discord_bp.from_config["client_id"] = "DISCORD_OAUTH_CLIENT_ID"