+++ tests/contrib/test_dropbox.py
@@ -1,5 +1,3 @@
-from __future__ import unicode_literals
-
 import pytest
 import responses
 from urlobject import URLObject
@@ -82,32 +80,39 @@ def test_context_local(make_app):
         assert request.headers["Authorization"] == "Bearer app2"
 
 
-def test_force_reapprove(make_app):
-    app = make_app("foo", "bar", force_reapprove=True)
-
+def app_redirect_location(app):
     with app.test_client() as client:
         resp = client.get("/dropbox", base_url="https://a.b.c", follow_redirects=False)
-    # check that there is a `force_reapprove=true` query param in the redirect URL
     assert resp.status_code == 302
-    location = URLObject(resp.headers["Location"])
-    assert location.query_dict["force_reapprove"] == "true"
+    return URLObject(resp.headers["Location"])
+
+
+def test_default_redirect_params(make_app):
+    app = make_app("foo", "bar")
+    query_dict = app_redirect_location(app).query_dict
+    assert isinstance(query_dict.pop("state"), str)
+    assert query_dict == {
+        "client_id": "foo",
+        "redirect_uri": "https://a.b.c/dropbox/authorized",
+        "response_type": "code",
+    }
+
+
+def test_force_reapprove(make_app):
+    app = make_app("foo", "bar", force_reapprove=True)
+    assert app_redirect_location(app).query_dict["force_reapprove"] == "true"
 
 
 def test_disable_signup(make_app):
     app = make_app("foo", "bar", disable_signup=True)
-
-    with app.test_client() as client:
-        resp = client.get("/dropbox", base_url="https://a.b.c", follow_redirects=False)
-    assert resp.status_code == 302
-    location = URLObject(resp.headers["Location"])
-    assert location.query_dict["disable_signup"] == "true"
+    assert app_redirect_location(app).query_dict["disable_signup"] == "true"
 
 
 def test_require_role(make_app):
     app = make_app("foo", "bar", require_role="work")
+    assert app_redirect_location(app).query_dict["require_role"] == "work"
 
-    with app.test_client() as client:
-        resp = client.get("/dropbox", base_url="https://a.b.c", follow_redirects=False)
-    assert resp.status_code == 302
-    location = URLObject(resp.headers["Location"])
-    assert location.query_dict["require_role"] == "work"
+
+def test_offline(make_app):
+    app = make_app("foo", "bar", offline=True)
+    assert app_redirect_location(app).query_dict["token_access_type"] == "offline"