+++ tests/contrib/test_twitch.py
@@ -2,19 +2,22 @@ import pytest
 import responses
 from urlobject import URLObject
 from flask import Flask
-from flask_dance.contrib.strava import make_strava_blueprint, strava
+from flask_dance.contrib.twitch import (
+    make_twitch_blueprint,
+    twitch,
+)
 from flask_dance.consumer import OAuth2ConsumerBlueprint
 from flask_dance.consumer.storage import MemoryStorage
 
 
 @pytest.fixture
 def make_app():
-    "A callable to create a Flask app with the  Strava provider"
+    "A callable to create a Flask app with the Twitch provider"
 
     def _make_app(*args, **kwargs):
         app = Flask(__name__)
         app.secret_key = "whatever"
-        blueprint = make_strava_blueprint(*args, **kwargs)
+        blueprint = make_twitch_blueprint(*args, **kwargs)
         app.register_blueprint(blueprint)
         return app
 
@@ -22,26 +25,33 @@ def make_app():
 
 
 def test_blueprint_factory():
-    strava_bp = make_strava_blueprint(
-        client_id="foo", client_secret="bar", scope="identity", redirect_to="index"
+    scope = [
+        "user:read:subscriptions",
+        "channel:read:subscriptions",
+        "user:read:subscriptions",
+    ]
+
+    twitch_bp = make_twitch_blueprint(
+        client_id="foo",
+        client_secret="bar",
+        scope=scope,
+        redirect_to="index",
     )
-    assert isinstance(strava_bp, OAuth2ConsumerBlueprint)
-    assert strava_bp.session.scope == "identity"
-    assert strava_bp.session.base_url == "https://www.strava.com/api/v3"
-    assert strava_bp.session.client_id == "foo"
-    assert strava_bp.client_secret == "bar"
-    assert (
-        strava_bp.authorization_url == "https://www.strava.com/api/v3/oauth/authorize"
-    )
-    assert strava_bp.token_url == "https://www.strava.com/api/v3/oauth/token"
+    assert isinstance(twitch_bp, OAuth2ConsumerBlueprint)
+    assert twitch_bp.session.scope == scope
+    assert twitch_bp.session.base_url == "https://api.twitch.tv/helix/"
+    assert twitch_bp.session.client_id == "foo"
+    assert twitch_bp.client_secret == "bar"
+    assert twitch_bp.authorization_url == "https://id.twitch.tv/oauth2/authorize"
+    assert twitch_bp.token_url == "https://id.twitch.tv/oauth2/token"
 
 
 def test_load_from_config(make_app):
     app = make_app()
-    app.config["STRAVA_OAUTH_CLIENT_ID"] = "foo"
-    app.config["STRAVA_OAUTH_CLIENT_SECRET"] = "bar"
+    app.config["TWITCH_OAUTH_CLIENT_ID"] = "foo"
+    app.config["TWITCH_OAUTH_CLIENT_SECRET"] = "bar"
 
-    resp = app.test_client().get("/strava")
+    resp = app.test_client().get("/twitch")
     url = resp.headers["Location"]
     client_id = URLObject(url).query.dict.get("client_id")
     assert client_id == "foo"
@@ -58,6 +68,7 @@ def test_context_local(make_app):
         redirect_to="url1",
         storage=MemoryStorage({"access_token": "app1"}),
     )
+
     app2 = make_app(
         "foo2",
         "bar2",
@@ -65,21 +76,23 @@ def test_context_local(make_app):
         storage=MemoryStorage({"access_token": "app2"}),
     )
 
-    # outside of a request context, referencing functions on the `strava` object
+    # outside of a request context, referencing functions on the `twitch` object
     # will raise an exception
     with pytest.raises(RuntimeError):
-        strava.get("https://google.com")
+        twitch.get("https://google.com")
 
-    # inside of a request context, `strava` should be a proxy to the correct
+    # inside of a request context, `twitch` should be a proxy to the correct
     # blueprint session
     with app1.test_request_context("/"):
         app1.preprocess_request()
-        strava.get("https://google.com")
+        twitch.get("https://google.com")
         request = responses.calls[0].request
         assert request.headers["Authorization"] == "Bearer app1"
+        assert request.headers["client-id"] == "foo1"
 
     with app2.test_request_context("/"):
         app2.preprocess_request()
-        strava.get("https://google.com")
+        twitch.get("https://google.com")
         request = responses.calls[1].request
         assert request.headers["Authorization"] == "Bearer app2"
+        assert request.headers["client-id"] == "foo2"